EDU-583 Extract module-utiles
